Refactor MesosSchedulerClient so that sink takes a SinkOperation
The new `SinkOperation` encapsulates the call to be sent to the server, with a potential `onComplete` and `onError`
handlers.

When `onError` is called an exception capturing the context of the request (MesosClientErrorContext) will be one of
the fields of the exception. There are different exceptions representing the different categories of errors that can
be encountered when communicating with the server. `MesosClientException` represents HTTP 400 class errors.
`MesosServerException` represents HTTP 500 class errors.

When a stream of `Observable<SinkOperation<Send>>` is passed to `sink` it is now subscribed to with a custom subscriber.
This pattern makes much more sense from an Rx and Event stream stand point as a Subscriber represents the thing
that consumes the events from the Observable.