Factor out `ConcatenatedInputStream` from `MessageStream`
This change explictly separates two different but related concerns:
aggregating incoming `ByteBuf`s for the `RecordIOOperator`, and
parsing RecordIO messages from the aggregated input.

The `ConcatenatedInputStream` class is a general-purpose class for
treating a queue of `InputStream`s as a single
`InputStream`. Additional streams can be added to the queue as they
become available.

The `MessageStream` class parses RecordIO messages from a single
`InputStream` provided in the constructor. The `InputStream` is
assumed to be backed by a source of bytes that can grow over time,
such that end-of-stream only means that all currently available data
has been consumed; subsequent attempts to read the stream may
succeed if new data has been received.

This change should make it easier in the future for `RecordIOOperator`
to support backpressure, as it allows the rate at which input
`ByteBuf`s are received to be decoupled from the rate at which parsed
output messages are sent.