Refactor MesosSchedulerClient API for improve lifecycle management.
Creation of a MesosSchedulerClient is now managed via the new
MesosSchedulerClientBuilder class. This class provides all the necessary
methods to construct a MesosSchedulerClient.

MesosSchedulerClient has been refactored to instead of returning a raw
`Observable<Protos.Event>` it instead task a
`Function<Observable<Protos.Event>, Observable<Optional<SinkOperation<Protos.Call>>>`.
This function will be woven into the event stream for the user. This
change affords us much greater control over the lifecycle of the event
stream and subscriptions on the end of it.

When the event stream is opened an `AwaitableSubscription` will be
returned to the user which can be used to await the end of the stream
(either error or complete). If the stream ends due to an error, the
error will be propagated to the thread blocked by the `await()` call.

Create a new utility TcpSocketProxy that can be used in tests. This
simple proxy allows for the creation of a proxy between two addresses.
The user can then close all sockets immediately or defer to a latter
point.