Add Mesos Scheduler Simulation
Create a new class MesosSchedulerSimulation which provides the ability
to run a simulated mesos master returning events provided to/via a
specified Observable<Protos.Event>.

As part of this new server a new module mesos-rxjava-util had to be
created to house some classes that are needed by mesos-rxjava-test. This
new module ensures that we don't have cyclic dependecies amongst our
modules.