Update Mesos dependency to 1.0.0-rc1 and protobuf to 2.6.1
Update Mesos dependency to use 1.0.0-rc1 which requires protobuf 2.6.1.

The decode test in ProtobufMessageCodecsTest have been removed in this update
because the new mesos protos have all properties marked as optional (no longer
some of them are required).  Since all properties are optional it is now
possible to "successfully" parse the bytes of a different message type into
the desired message (the bytes of a heartbeat can be successfully parsed as a
Subscribe call).