Update MesosClient to decode error response body on subscribe
MesosClient will new attempt to decode the response body from mesos in
the case that the response status to the initial subscribe call is
non-200.

If the response from mesos has a content type of text/plain the response
body will be read as a utf-8 string and included as the message in the
MesosClientErrorContext in the exception throw.

Fixes #66