Update non 202 response handling for consistency
Update non 202 response handling of SinkSubscriber so that it follows
the same pattern MesosClient uses when subscribe calls. Primarily
regarding how the response body is handled. (If the response is
'text/plain' the body will be read as a UTF-8 string and set as the
message property of the MesosClientErrorContext in the exception).