Retain auth information after redirects
When passed a user:password to connect to mesos as part of the master uri,
that auth info is not retained when redirected to a new master. This updates
the client to retain the original auth info provided when the
original master we connect to is not the leader and we get redirected.